[FIX] l10n_ve_stock_account,l10n_ve_donation: invierte dependencia de donación y corrige desinstalación
Invierte la dependencia para que l10n_ve_donation dependa de

l10n_ve_stock_account, en lugar del sentido anterior. Esto evita el

error KeyError al desinstalar el módulo de donaciones cuando otros

módulos referencian sus campos.

- Elimina toda la lógica de donación (campos, métodos, vistas,

  traducciones y reportes) de l10n_ve_stock_account

- Mueve la lógica de stock.picking a l10n_ve_donation con herencia

  limpia mediante super()

- Crea stock_picking_views.xml en donation heredando de

  stock_account

- Reemplaza _prepare_invoice_values (código muerto) por override real

  de create_invoice que propaga is_donation a las facturas

- Elimina _get_tax_grouped_lines duplicado en donation (usa el de

  stock_account por MRO)

- Cambia auto_install a False en donation para prevenir colisiones

  automáticas

- Limpia entradas de traducción de donación del .po de stock_account

- Elimina métodos vacíos action_post y _onchange_partner_id de

  stock_account

References: https://www.binauraldev.com/odoo/helpdesk/58/tickets/13060
