+++ l10n_ve_donation/models/__init__.py
@@ -7,6 +7,7 @@ from . import (
     sale_order,
     stock_location,
     stock_move,
+    stock_picking,
     stock_picking_type,
     stock_scrap,
     stock_warehouse,
