+++ l10n_ve_donation/views/stock_picking_views.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="stock_picking_form_view_donation" model="ir.ui.view">
+        <field name="name">stock.picking.form.view.donation</field>
+        <field name="model">stock.picking</field>
+        <field name="inherit_id" ref="l10n_ve_stock_account.stock_picking_form_view_inherited" />
+        <field name="arch" type="xml">
+            <!-- Add is_donation field in form (visible for editing) -->
+            <xpath expr="//field[@name='partner_id']" position="before">
+                <field name="is_donation" readonly="is_donation" force_save="1"/>
+            </xpath>
+
+            <!-- Make partner_id readonly when is_donation -->
+            <xpath expr="//form//group//group//field[@name='partner_id']" position="attributes">
+                <attribute name="readonly">order_is_consignment or partner_required or is_donation</attribute>
+            </xpath>
+
+            <!-- Make outgoing transfer_reason_id readonly when is_donation -->
+            <xpath expr="//field[@name='transfer_reason_id'][not(@required)]" position="attributes">
+                <attribute name="readonly">is_donation or state == 'done'</attribute>
+            </xpath>
+
+            <!-- Hide guide_number when is_donation and no dispatch controls -->
+            <xpath expr="//field[@name='guide_number']" position="attributes">
+                <attribute name="invisible">not dispatch_guide_controls and is_donation</attribute>
+            </xpath>
+        </field>
+    </record>
+</odoo>
