+++ l10n_ve_stock_account/__manifest__.py
@@ -7,13 +7,12 @@
     "author": "binaural-dev",
     "website": "https://www.binauraldev.com",
     "category": "Stock Account",
-    "version": "19.0.1.1.4",
+    "version": "19.0.2.0.0",
     "depends": [
         "l10n_ve_stock",
         "l10n_ve_invoice",
         "l10n_ve_accountant",
         "l10n_ve_sale",
-        "l10n_ve_donation",
         "sale_stock",
         "web",
     ],
