+++ l10n_ve_stock_account/views/sale_order_views.xml
@@ -2,16 +2,13 @@
     <record id="view_order_form_l10n_ve_stock_account" model="ir.ui.view">
         <field name="name">sale.order.form.l10n.ve.stock.account</field>
         <field name="model">sale.order</field>
-        <field name="inherit_id" ref="l10n_ve_donation.view_order_form_l10n_ve_donation" />
+        <field name="inherit_id" ref="sale.view_order_form" />
         <field name="arch" type="xml">
-            <field name="payment_term_id" position="after">
+            <xpath expr="//field[@name='payment_term_id']" position="after">
                 <field name="is_consignation" invisible="1"/>
-                <field name="document" readonly="state in ['sale', 'done', 'cancel'] or is_consignation or is_donation" />
-            </field>
-            <field name="is_donation" position="replace">
-                <field name="is_donation" readonly="state in ['sale', 'done', 'cancel']" force_save="1"/>
+                <field name="document" readonly="state in ['sale', 'done', 'cancel'] or is_consignation" />
                 <field name="is_consignation" readonly="state in ['sale', 'done', 'cancel']" />
-            </field>
+            </xpath>
             <xpath expr="//button[@id='create_invoice']" position="attributes">
                 <attribute name="invisible">(invoice_status != 'to invoice' and invoice_status != 'partially_billed') or document == 'dispatch_guide'</attribute>
             </xpath>
